A fun lottery game that you will never win
   Lottery Game

This Program is used to guess a set of 3 numbers that were chosen by the computer. The computer uses a complex algorithm to randomly change the number for each trial. The game is called lottery game, it means the lucky one who guess one or more numbers is rewarded an amount proportional to the number of numbers where found. For this program, if the user guesses 1 number among the 32 chosen by the computer, the user will have won $1,000.  For 2 numbers guessed, the user has $5,000. The Jackpot is $10,000 when the user has entered all three numbers that the computer has chosen.

This Lottery game does not consider the numbers in order. This means that it doesn’t matter whether the first number is the last, as long as it matches. This Program is based on If and else if statements for its matching of numbers. This Program is complex and requires coordination for the values entered by the user. However, this Program is also very impressive regarding the changing of the number chosen by the computer every time the user plays the game. This Game can be upgraded and could be inserted in a lottery machine.