Rock Paper Scissors
This Program is a game between the user and the computer. This game is really famous. It is called ROCK Paper and Scissors. Many of us have already played this game either for pleasure or for passing time. Here we are asked to make this game so that the player can enter his choice and the computer chooses randomly his choice. The Program compares the two choices and makes a decision that follows an algorithm that I have designed for the program.

Basically, this game is developed by using a function called RAND which makes the computer chose between 3 numbers every time we try. The choice is random. Following the rules of the Rock Paper Scissors game, we already know that the paper beats the rock the rock beats the scissors and the scissors beat the paper. The action behind this program is to handle all the cases that can happen and assign an action according to it.